+++ lib/bambora/bank/payment_profile_resource.rb
@@ -18,12 +18,12 @@ module Bambora
       #
       # @example
       #
-      #   client = Bambora::WWWFormClient(base_url: '...', merchant_id: '...')
+      #   client = Bambora::Rest::WWWFormClient(base_url: '...', merchant_id: '...')
       #   profiles = Bambora::Bank::PaymentProfileResource(client: client, api_key: '...')
       #
       #   # Start making requests ...
       #
-      # @param client [Bambora::WWWFormClient] An instance of Bambora::WWWFormClient, used to make network requests.
+      # @param client [Bambora::Rest::WWWFormClient] An instance of Bambora::Rest::WWWFormClient, used to make network requests.
       # @param api_key [String] An API key for this endpoint. This is also known as the "Pass Code"
       # @param version [String] The Service Version you are requesting from the server.
       def initialize(client:, api_key:, version: DEFAULT_VERSION)
